Struct chemfiles::Frame

pub struct Frame { /* fields omitted */ }

A Frame holds data from one step of a simulation: the current Topology, the positions, and maybe the velocities of the particles in the system.

Methods

impl Frame

fn new(natoms: usize) -> Result<Frame>

Create an empty frame with initial capacity of natoms. It will be resized by the library as needed.

fn atom(&self, index: usize) -> Result<Atom>

Get a specific Atom from a frame, given its index in the frame

fn natoms(&self) -> Result<usize>

Get the current number of atoms in the Frame.

fn resize(&mut self, natoms: usize) -> Result<()>

Resize the positions and the velocities in frame, to make space for natoms atoms. Previous data is conserved, as well as the presence of absence of velocities.

fn positions(&self) -> Result<&[[f32; 3]]>

Get a view into the positions of the Frame.

fn positions_mut(&mut self) -> Result<&mut [[f32; 3]]>

Get a mutable view into the positions of the Frame.

fn velocities(&self) -> Result<&[[f32; 3]]>

Get a view into the velocities of the Frame.

fn velocities_mut(&mut self) -> Result<&mut [[f32; 3]]>

Get a mutable view into the velocities of the Frame.

fn has_velocities(&self) -> Result<bool>

Check if the Frame has velocity information.

fn add_velocities(&mut self) -> Result<()>

Add velocity storage to this frame for Frame::natoms atoms. If the frame already have velocities, this does nothing.

fn cell(&self) -> Result<UnitCell>

Get the UnitCell from the Frame

fn set_cell(&mut self, cell: &UnitCell) -> Result<()>

Set the UnitCell of the Frame

fn topology(&self) -> Result<Topology>

Get the Topology from the Frame

fn set_topology(&mut self, topology: &Topology) -> Result<()>

Set the Topology of the Frame

fn step(&self) -> Result<usize>

Get the Frame step, i.e. the frame number in the trajectory

fn set_step(&mut self, step: usize) -> Result<()>

Set the Frame step

fn guess_topology(&mut self) -> Result<()>

Guess the bonds, angles and dihedrals in the system using an atomic distance criteria

unsafe fn from_ptr(ptr: *const CHFL_FRAME) -> Frame

Create a Frame from a C pointer. This function is unsafe because no validity check is made on the pointer.

unsafe fn as_ptr(&self) -> *const CHFL_FRAME

Get the underlying C pointer. This function is unsafe because no lifetime guarantee is made on the pointer.