Struct chemfiles::Logger

pub struct Logger<'a> {
    // some fields omitted
}

This struct give access to the logging system.

As it is a global system, it must be aquired before any operations.

Methods

impl<'a> Logger<'a>

fn get() -> Logger<'a>

Get an handle to the logging system. This function blocks, waiting for a mutex to be available. You should probably call this function from one thread only.

fn level(&self) -> Result<LogLevel, Error>

Get the current maximal logging level

fn set_level(&self, level: LogLevel) -> Result<(), Error>

Set the maximal logging level to level

fn log_to_file<P>(&self, filename: P) -> Result<(), Error> where P: AsRef<Path>

Write logs to the file at path, creating it if needed.

fn log_to_stderr(&self) -> Result<(), Error>

Redirect the logs to the standard error stream. This is the default.

fn log_to_stdout(&self) -> Result<(), Error>

Redirect the logs to the standard output.

fn log_silent(&self) -> Result<(), Error>

Remove all logging output.

fn log_callback<F>(&self, callback: F) -> Result<(), Error> where F: Fn(LogLevel, &str) + 'static

Redirect all logging to user-provided logging. The callback function will be called at each loggin operation with the level of the message, and the the message itself.