Struct chemfiles::Selection

pub struct Selection { /* fields omitted */ }

A Selection allow to select atoms in a Frame, from a selection language. The selection language is built by combining basic operations. Each basic operation follows the <selector>[(<variable>)] <operator> <value> structure, where <operator> is a comparison operator in == != < <= > >=.

Methods

impl Selection

pub fn new<'a, S: Into<&'a str>>(selection: S) -> Result<Selection>

Create a new selection from the given selection string.

Example

let selection = Selection::new("pairs: name(#1) H and name(#2) O").unwrap();

pub fn size(&self) -> Result<u64>

Get the size of the selection, i.e. the number of atoms we are selecting together.

This value is 1 for the 'atom' context, 2 for the 'pair' and 'bond' context, 3 for the 'three' and 'angles' contextes and 4 for the 'four' and 'dihedral' contextes.

Example

let selection = Selection::new("pairs: name(#1) H and name(#2) O").unwrap();
assert_eq!(selection.size(), Ok(2));

pub fn string(&self) -> Result<String>

Get the selection string used to create this selection.

Example

let selection = Selection::new("name H").unwrap();
assert_eq!(selection.string(), Ok(String::from("name H")));

pub fn evaluate(&mut self, frame: &Frame) -> Result<Vec<Match>>

Evaluate a selection for a given frame, and return the corresponding matches.

Example

let mut frame = Frame::new().unwrap();
frame.add_atom(&Atom::new("H").unwrap(), [1.0, 0.0, 0.0], None).unwrap();
frame.add_atom(&Atom::new("O").unwrap(), [0.0, 0.0, 0.0], None).unwrap();
frame.add_atom(&Atom::new("H").unwrap(), [-1.0, 0.0, 0.0], None).unwrap();

let mut selection = Selection::new("pairs: name(#1) H and name(#2) O").unwrap();
let matches = selection.evaluate(&frame).unwrap();

assert_eq!(matches.len(), 2);

assert_eq!(matches[0].len(), 2);
assert_eq!(matches[0][0], 0);
assert_eq!(matches[0][1], 1);

assert_eq!(matches[1].len(), 2);
assert_eq!(matches[1][0], 2);
assert_eq!(matches[1][1], 1);

pub fn list(&mut self, frame: &Frame) -> Result<Vec<u64>>

Evaluates a selection of size 1 on a given frame. This function returns the list of atomic indexes in the frame matching this selection.

Panics

If the selection size is not 1.

Example

let mut frame = Frame::new().unwrap();
frame.add_atom(&Atom::new("H").unwrap(), [1.0, 0.0, 0.0], None).unwrap();
frame.add_atom(&Atom::new("O").unwrap(), [0.0, 0.0, 0.0], None).unwrap();
frame.add_atom(&Atom::new("H").unwrap(), [-1.0, 0.0, 0.0], None).unwrap();

let mut selection = Selection::new("name H").unwrap();
let matches = selection.list(&frame).unwrap();

assert_eq!(matches.len(), 2);
assert_eq!(matches[0], 0);
assert_eq!(matches[1], 2);