Struct chemfiles::Trajectory

pub struct Trajectory { /* fields omitted */ }

A Trajectory is a chemistry file on the hard drive. It is the main entry point of chemfiles.

Methods

impl Trajectory

fn open<P>(filename: P) -> Result<Trajectory> where P: AsRef<Path>

Open a trajectory file in read mode.

fn create<P>(filename: P) -> Result<Trajectory> where P: AsRef<Path>

Open a trajectory file in write mode.

fn open_with_format<'a, P, S>(filename: P, format: S) -> Result<Trajectory> where P: AsRef<Path>, S: Into<&'a str>

Open a trajectory file in read mode using a specific format.

fn create_with_format<'a, P, S>(filename: P, format: S) -> Result<Trajectory> where P: AsRef<Path>, S: Into<&'a str>

Open a trajectory file in write mode.

fn read(&mut self, frame: &mut Frame) -> Result<()>

Read the next step of the trajectory into a frame

fn read_step(&mut self, step: usize, frame: &mut Frame) -> Result<()>

Read a specific step of the trajectory in a frame

fn write(&mut self, frame: &Frame) -> Result<()>

Write a frame to the trajectory.

fn set_topology(&mut self, topology: Topology) -> Result<()>

Set the topology associated with a trajectory. This topology will be used when reading and writing the files, replacing any topology in the frames or files.

fn set_topology_file<P>(&mut self, filename: P) -> Result<()> where P: AsRef<Path>

Set the topology associated with a trajectory by reading the first frame of filename; and extracting the topology of this frame.

fn set_cell(&mut self, cell: UnitCell) -> Result<()>

Set the unit cell associated with a trajectory. This cell will be used when reading and writing the files, replacing any unit cell in the frames or files.

fn nsteps(&mut self) -> Result<usize>

Get the number of steps (the number of frames) in a trajectory.

fn sync(&mut self) -> Result<()>

Synchronize any buffered content to the hard drive.

unsafe fn from_ptr(ptr: *mut CHFL_TRAJECTORY) -> Trajectory

Create a Trajectory from a C pointer. This function is unsafe because no validity check is made on the pointer.

unsafe fn as_ptr(&self) -> *const CHFL_TRAJECTORY

Get the underlying C pointer. This function is unsafe because no lifetime guarantee is made on the pointer.