Crate chemfiles

Chemfiles is a multi-language library written in modern C++ for reading and writing from and to molecular trajectory files. These files are created by your favorite theoretical chemistry program, and contains informations about atomic or residues names and positions. Some format also have additional informations, such as velocities, forces, energy, …

This crate expose the C API of chemfiles to Rust, and make all the functionalities accessibles. For more informations on the C++ library, please see its documentation. Specifically, the following pages are worth reading:

The overview of the classes organisation;

The supported formats;

As all the function call the underlying C library, they all can fail and thus all return a Result<_, Error> value.

Structs

Atom
An Atom is a particle in the current Frame. It can be used to store and retrieve informations about a particle, such as mass, name, atomic number, etc.

Error
Error type for Chemfiles.

Frame
A Frame holds data from one step of a simulation: the current Topology, the positions, and maybe the velocities of the particles in the system.

Logger
This struct give access to the logging system.

Topology
A Topology contains the definition of all the particles in the system, and the liaisons between the particles (bonds, angles, dihedrals, ...).

Trajectory
A Trajectory is a chemistry file on the hard drive. It is the main entry point of chemfiles.

UnitCell
An UnitCell represent the box containing the atoms in the system, and its periodicity.

Enums

AtomType
Available types of atoms

CellType
Available unit cell types

Functions

version
Get the version of the chemfiles library