Function chemfiles::add_configuration

pub fn add_configuration<S>(path: S) -> Result<()> where
    S: AsRef<str>, 

Read configuration data from the file at path.

By default, chemfiles reads configuration from any file name .chemfilesrc in the current directory or any parent directory. This function can be used to add data from another configuration file.

This function will fail if there is no file at path, or if the file is incorectly formatted. Data from the new configuration file will overwrite any existing data.

Example

chemfiles::add_configuration("local-config.toml");
// from now on, the data from "local-config.toml" will be used