Struct chemfiles::Atom

pub struct Atom {
    // some fields omitted
}

An Atom is a particle in the current Frame. It can be used to store and retrieve informations about a particle, such as mass, name, atomic number, etc.

Methods

impl Atom

fn new<'a, S>(name: S) -> Result<Atom, Error> where S: Into<&'a str>

Create a new Atom from a name.

fn mass(&self) -> Result<f32, Error>

Get the Atom mass, in atomic mass units

fn set_mass(&mut self, mass: f32) -> Result<(), Error>

Set the Atom mass, in atomic mass units

fn charge(&self) -> Result<f32, Error>

Get the Atom charge, in number of the electron charge e

fn set_charge(&mut self, charge: f32) -> Result<(), Error>

Set the Atom charge, in number of the electron charge e

fn name(&self) -> Result<String, Error>

Get the Atom name

fn set_name<'a, S>(&mut self, name: S) -> Result<(), Error> where S: Into<&'a str>

Set the Atom name

fn full_name(&mut self) -> Result<String, Error>

Try to get the full name of the Atom. The full name of "He" is "Helium", and so on. If the name can not be found, returns the empty string.

fn vdw_radius(&self) -> Result<f64, Error>

Try to get the Van der Waals radius of the Atom. If the radius can not be found, returns -1.

fn covalent_radius(&self) -> Result<f64, Error>

Try to get the covalent radius of the Atom. If the radius can not be found, returns -1.

fn atomic_number(&self) -> Result<i32, Error>

Try to get the atomic number of the Atom. If the number can not be found, returns -1.

fn atom_type(&self) -> Result<AtomType, Error>

Get the type of the atom

fn set_atom_type(&mut self, atom_type: AtomType) -> Result<(), Error>

Set the type of the atom

unsafe fn from_ptr(ptr: *const CHFL_ATOM) -> Atom

Create an Atom from a C pointer. This function is unsafe because no validity check is made on the pointer.

unsafe fn as_ptr(&self) -> *const CHFL_ATOM

Get the underlying C pointer. This function is unsafe because no lifetime guarantee is made on the pointer.