Struct chemfiles::Frame

pub struct Frame {
    // some fields omitted
}

A Frame holds data from one step of a simulation: the current Topology, the positions, and maybe the velocities of the particles in the system.

Methods

impl Frame

fn new(natoms: usize) -> Result<Frame, Error>

Create an empty frame with initial capacity of natoms. It will be resized by the library as needed.

fn atom(&self, index: usize) -> Result<Atom, Error>

Get a specific Atom from a frame, given its index in the frame

fn natoms(&self) -> Result<usize, Error>

Get the current number of atoms in the Frame.

fn positions(&self) -> Result<Vec<[f32; 3]>, Error>

Get the positions from the Frame.

fn set_positions(&mut self, positions: Vec<[f32; 3]>) -> Result<(), Error>

Set the positions in the Frame.

fn velocities(&self) -> Result<Vec<[f32; 3]>, Error>

Get the velocities from the Frame.

fn set_velocities(&mut self, velocities: Vec<[f32; 3]>) -> Result<(), Error>

Set the velocities in the Frame.

fn has_velocities(&self) -> Result<bool, Error>

Check if the Frame has velocity information.

fn cell(&self) -> Result<UnitCell, Error>

Get the UnitCell from the Frame

fn set_cell(&mut self, cell: &UnitCell) -> Result<(), Error>

Set the UnitCell of the Frame

fn topology(&self) -> Result<Topology, Error>

Get the Topology from the Frame

fn set_topology(&mut self, topology: &Topology) -> Result<(), Error>

Set the Topology of the Frame

fn step(&self) -> Result<usize, Error>

Get the Frame step, i.e. the frame number in the trajectory

fn set_step(&mut self, step: usize) -> Result<(), Error>

Set the Frame step

fn guess_topology(&self, bonds: bool) -> Result<(), Error>

Try to guess the bonds, angles and dihedrals in the system. If bonds is true, guess everything; else only guess the angles and dihedrals from the topology bond list.

unsafe fn from_ptr(ptr: *const CHFL_FRAME) -> Frame

Create a Frame from a C pointer. This function is unsafe because no validity check is made on the pointer.

unsafe fn as_ptr(&self) -> *const CHFL_FRAME

Get the underlying C pointer. This function is unsafe because no lifetime guarantee is made on the pointer.