Struct chemfiles::Residue

pub struct Residue { /* fields omitted */ }

A Residue is a group of atoms belonging to the same logical unit. They can be small molecules, amino-acids in a protein, monomers in polymers, etc.

Implementations

impl Residue

pub fn new<'a>(name: impl Into<&'a str>) -> Residue

Create a new residue with the given name

Example

let residue = Residue::new("ALA");
assert_eq!(residue.name(), "ALA");
assert_eq!(residue.id(), None);

pub fn with_id<'a>(name: impl Into<&'a str>, id: i64) -> Residue

Create a new residue with the given name and id as identifier.

Example

let residue = Residue::with_id("ALA", 67);
assert_eq!(residue.name(), "ALA");
assert_eq!(residue.id(), Some(67));

pub fn size(&self) -> usize

Get the number of atoms in this residue.

Example

let mut residue = Residue::new("water");
assert_eq!(residue.size(), 0);

residue.add_atom(0);
residue.add_atom(1);
residue.add_atom(2);
assert_eq!(residue.size(), 3);

pub fn id(&self) -> Option<i64>

Get the identifier of this residue in the initial topology file.

Example

let residue = Residue::with_id("", 42);
assert_eq!(residue.id(), Some(42));

pub fn name(&self) -> String

Get the name of this residue.

Example

let residue = Residue::new("water");
assert_eq!(residue.name(), "water");

pub fn add_atom(&mut self, atom: usize)

Add the atom at index atom in this residue.

This will fail if the atom is already in the residue.

Example

let mut residue = Residue::new("water");
assert_eq!(residue.size(), 0);
assert_eq!(residue.contains(56), false);

residue.add_atom(56);
assert_eq!(residue.size(), 1);
assert_eq!(residue.contains(56), true);

// Adding the same atom twice is fine
residue.add_atom(56);
assert_eq!(residue.size(), 1);

pub fn contains(&self, atom: usize) -> bool

Check if the atom at index i is in this residue

Example

let mut residue = Residue::new("water");
assert_eq!(residue.contains(56), false);

residue.add_atom(56);
assert_eq!(residue.contains(56), true);

pub fn atoms(&self) -> Vec<usize>

Get the list of atoms of this residue.

Example

let mut residue = Residue::new("water");
assert_eq!(residue.atoms(), vec![]);

residue.add_atom(56);
assert_eq!(residue.atoms(), vec![56]);

pub fn set(&mut self, name: &str, property: impl Into<Property>)

Add a new property with the given name to this residue.

If a property with the same name already exists, this function override the existing property with the new one.

Examples

let mut residue = Residue::new("ALA");
residue.set("a string", "hello");
residue.set("a double", 3.2);

assert_eq!(residue.get("a string"), Some(Property::String("hello".into())));
assert_eq!(residue.get("a double"), Some(Property::Double(3.2)));

pub fn get(&self, name: &str) -> Option<Property>

Get a property with the given name in this frame, if it exist.

Examples

let mut residue = Residue::new("ALA");
residue.set("foo", Property::Double(22.2));

assert_eq!(residue.get("foo"), Some(Property::Double(22.2)));
assert_eq!(residue.get("Bar"), None);

pub fn properties(&self) -> PropertiesIter<'_>

Get an iterator over all (name, property) pairs for this frame

Examples

let mut residue = Residue::new("ALA");
residue.set("foo", Property::Double(22.2));
residue.set("bar", Property::Bool(false));

for (name, property) in residue.properties() {
    if name == "foo" {
        assert_eq!(property, Property::Double(22.2));
    } else if name == "bar" {
        assert_eq!(property, Property::Bool(false));
    }
}