Struct chemfiles::Selection

pub struct Selection {
    // some fields omitted
}

Select atoms in a Frame with a selection language.

The selection language is built by combining basic operations. Each basic operation follows the <selector>[(<variable>)] <operator> <value> structure, where <operator> is a comparison operator in == != < <= > >=. Refer to the full documentation to know the allowed selectors and how to use them.

Methods

impl Selection

fn new<'a, S: Into<&'a str>>(selection: S) -> Result<Selection>

Create a new selection from the given selection string.

fn size(&self) -> Result<usize>

Get the size of the selection, i.e. the number of atoms we are selecting together.

This value is 1 for the 'atom' context, 2 for the 'pair' and 'bond' context, 3 for the 'three' and 'angles' contextes and 4 for the 'four' and 'dihedral' contextes.

fn evaluate(&mut self, frame: &Frame) -> Result<Vec<Match>>

Evaluate a selection for a given frame, and return the corresponding matches.

fn list(&mut self, frame: &Frame) -> Result<Vec<usize>>

Evaluates a selection of size 1 on a given frame. This function returns the list of atomic indexes in the frame matching this selection.

Panics

If the selection size is not 1.