Struct chemfiles::Topology

pub struct Topology { /* fields omitted */ }

A Topology contains the definition of all the atoms in the system, and the liaisons between the atoms (bonds, angles, dihedrals, ...). It will also contain all the residues information if it is available.

Methods

impl Topology

fn new() -> Result<Topology>

Create a new empty topology.

Example

let topology = Topology::new().unwrap();
assert_eq!(topology.natoms(), Ok(0));

fn atom(&self, index: u64) -> Result<Atom>

Get a copy of the atom at index index from this topology.

Example

let mut topology = Topology::new().unwrap();
topology.resize(6).unwrap();

let atom = topology.atom(4).unwrap();
assert_eq!(atom.name(), Ok(String::new()));

fn natoms(&self) -> Result<u64>

Get the current number of atoms in this topology.

Example

let mut topology = Topology::new().unwrap();
assert_eq!(topology.natoms(), Ok(0));

topology.resize(6).unwrap();
assert_eq!(topology.natoms(), Ok(6));

fn resize(&mut self, natoms: u64) -> Result<()>

Resize this topology to hold natoms atoms, inserting dummy atoms if the new size if bigger than the old one.

Example

let mut topology = Topology::new().unwrap();
assert_eq!(topology.natoms(), Ok(0));

topology.resize(6).unwrap();
assert_eq!(topology.natoms(), Ok(6));

fn add_atom(&mut self, atom: &Atom) -> Result<()>

Add an Atom at the end of this topology

Example

let mut topology = Topology::new().unwrap();
topology.add_atom(&Atom::new("Mg").unwrap()).unwrap();

let atom = topology.atom(0).unwrap();
assert_eq!(atom.name(), Ok(String::from("Mg")));

fn remove(&mut self, index: u64) -> Result<()>

Remove an Atom from this topology by index. This modify all the other atoms indexes.

Example

let mut topology = Topology::new().unwrap();
topology.resize(9).unwrap();
assert_eq!(topology.natoms(), Ok(9));

topology.remove(7).unwrap();
assert_eq!(topology.natoms(), Ok(8));

fn is_bond(&self, i: u64, j: u64) -> Result<bool>

Tell if the atoms at indexes i and j are bonded together.

Example

let mut topology = Topology::new().unwrap();
assert_eq!(topology.is_bond(9, 8), Ok(false));

topology.add_bond(9, 8).unwrap();
assert_eq!(topology.is_bond(9, 8), Ok(true));

fn is_angle(&self, i: u64, j: u64, k: u64) -> Result<bool>

Tell if the atoms at indexes i, j and k constitues an angle.

Example

let mut topology = Topology::new().unwrap();
assert_eq!(topology.is_angle(9, 8, 6), Ok(false));

topology.add_bond(9, 8).unwrap();
topology.add_bond(6, 8).unwrap();
assert_eq!(topology.is_angle(9, 8, 6), Ok(true));

fn is_dihedral(&self, i: u64, j: u64, k: u64, m: u64) -> Result<bool>

Tell if the atoms at indexes i, j, k and m constitues a dihedral angle.

Example

let mut topology = Topology::new().unwrap();
assert_eq!(topology.is_dihedral(2, 9, 8, 6), Ok(false));

topology.add_bond(2, 9).unwrap();
topology.add_bond(9, 8).unwrap();
topology.add_bond(6, 8).unwrap();
assert_eq!(topology.is_dihedral(2, 9, 8, 6), Ok(true));

fn bonds_count(&self) -> Result<u64>

Get the number of bonds in the topology.

Example

let mut topology = Topology::new().unwrap();
assert_eq!(topology.bonds_count(), Ok(0));

topology.add_bond(2, 9).unwrap();
topology.add_bond(9, 8).unwrap();
topology.add_bond(6, 8).unwrap();
assert_eq!(topology.bonds_count(), Ok(3));

fn angles_count(&self) -> Result<u64>

Get the number of angles in the topology.

Example

let mut topology = Topology::new().unwrap();
assert_eq!(topology.angles_count(), Ok(0));

topology.add_bond(2, 9).unwrap();
topology.add_bond(9, 8).unwrap();
topology.add_bond(6, 8).unwrap();
assert_eq!(topology.angles_count(), Ok(2));

fn dihedrals_count(&self) -> Result<u64>

Get the number of dihedral angles in the topology.

Example

let mut topology = Topology::new().unwrap();
assert_eq!(topology.dihedrals_count(), Ok(0));

topology.add_bond(2, 9).unwrap();
topology.add_bond(9, 8).unwrap();
topology.add_bond(6, 8).unwrap();
assert_eq!(topology.dihedrals_count(), Ok(1));

fn bonds(&self) -> Result<Vec<[u64; 2]>>

Get the list of bonds in the topology.

Example

let mut topology = Topology::new().unwrap();
topology.add_bond(2, 9).unwrap();
topology.add_bond(9, 8).unwrap();
topology.add_bond(6, 8).unwrap();

assert_eq!(topology.bonds(), Ok(vec![[2, 9], [6, 8], [8, 9]]));

fn angles(&self) -> Result<Vec<[u64; 3]>>

Get the list of angles in the topology.

Example

let mut topology = Topology::new().unwrap();
topology.add_bond(2, 9).unwrap();
topology.add_bond(9, 8).unwrap();
topology.add_bond(6, 8).unwrap();

assert_eq!(topology.angles(), Ok(vec![[2, 9, 8], [6, 8, 9]]));

fn dihedrals(&self) -> Result<Vec<[u64; 4]>>

Get the list of dihedral angles in the topology.

Example

let mut topology = Topology::new().unwrap();
topology.add_bond(2, 9).unwrap();
topology.add_bond(9, 8).unwrap();
topology.add_bond(6, 8).unwrap();

assert_eq!(topology.dihedrals(), Ok(vec![[6, 8, 9, 2]]));

fn add_bond(&mut self, i: u64, j: u64) -> Result<()>

Add a bond between the atoms at indexes i and j in the topology.

Example

let mut topology = Topology::new().unwrap();
assert_eq!(topology.bonds_count(), Ok(0));

topology.add_bond(0, 1).unwrap();
topology.add_bond(9, 8).unwrap();
assert_eq!(topology.bonds_count(), Ok(2));

fn remove_bond(&mut self, i: u64, j: u64) -> Result<()>

Remove any existing bond between the atoms at indexes i and j in this topology.

This function does nothing if there is no bond between i and j.

Example

let mut topology = Topology::new().unwrap();
assert_eq!(topology.bonds_count(), Ok(0));

topology.add_bond(0, 1).unwrap();
topology.add_bond(9, 8).unwrap();
assert_eq!(topology.bonds_count(), Ok(2));

topology.remove_bond(0, 1).unwrap();
assert_eq!(topology.bonds_count(), Ok(1));

topology.remove_bond(10, 1).unwrap();
assert_eq!(topology.bonds_count(), Ok(1));

fn residue(&self, index: u64) -> Result<Residue>

Get a copy of the residue at index index from this topology.

The residue index in the topology is not always the same as the residue id.

Example

let mut topology = Topology::new().unwrap();
topology.add_residue(&Residue::new("water").unwrap()).unwrap();

let residue = topology.residue(0).unwrap();
assert_eq!(residue.name(), Ok(String::from("water")));

fn residue_for_atom(&self, index: u64) -> Result<Option<Residue>>

Get a copy of the residue containing the atom at index index in this topology, if any.

Example

let mut topology = Topology::new().unwrap();
topology.resize(8).unwrap();

let mut residue = Residue::new("water").unwrap();
residue.add_atom(0).unwrap();
residue.add_atom(1).unwrap();
residue.add_atom(2).unwrap();
topology.add_residue(&residue).unwrap();

let residue = topology.residue_for_atom(0).unwrap().unwrap();
assert_eq!(residue.name(), Ok(String::from("water")));

let residue = topology.residue_for_atom(6).unwrap();
assert!(residue.is_none());

fn residues_count(&self) -> Result<u64>

Get the number of residues in this topology.

Example

let mut topology = Topology::new().unwrap();
assert_eq!(topology.residues_count(), Ok(0));

topology.add_residue(&Residue::with_id("water", 0).unwrap()).unwrap();
topology.add_residue(&Residue::with_id("protein", 1).unwrap()).unwrap();
assert_eq!(topology.residues_count(), Ok(2));

fn add_residue(&mut self, residue: &Residue) -> Result<()>

Add a residue to this topology.

The residue id must not already be in the topology, and the residue must contain only atoms that are not already in another residue.

Example

let mut topology = Topology::new().unwrap();
topology.add_residue(&Residue::new("water").unwrap()).unwrap();

let residue = topology.residue(0).unwrap();
assert_eq!(residue.name(), Ok(String::from("water")));

fn are_linked(&self, first: &Residue, second: &Residue) -> Result<bool>

Check if the two residues first and second from the topology are linked together, i.e. if there is a bond between one atom in the first residue and one atom in the second one.

Example

let mut topology = Topology::new().unwrap();

topology.add_residue(&Residue::with_id("water", 0).unwrap()).unwrap();
topology.add_residue(&Residue::with_id("protein", 1).unwrap()).unwrap();

let first = topology.residue(0).unwrap();
let second = topology.residue(1).unwrap();
assert_eq!(topology.are_linked(&first, &second), Ok(false));