Struct chemfiles::Topology

pub struct Topology {
    // some fields omitted
}

A Topology contains the definition of all the particles in the system, and the liaisons between the particles (bonds, angles, dihedrals, ...).

Only the atoms and the bonds are stored, the angles and the dihedrals are computed automaticaly.

Methods

impl Topology

fn new() -> Result<Topology, Error>

Create a new empty topology

fn atom(&self, index: usize) -> Result<Atom, Error>

Get a specific Atom from a topology, given its index in the topology

fn natoms(&self) -> Result<usize, Error>

Get the current number of atoms in the topology.

fn push(&mut self, atom: &Atom) -> Result<(), Error>

Add an Atom at the end of a topology

fn remove(&mut self, index: usize) -> Result<(), Error>

Remove an Atom from a topology by index. This modify all the other atoms indexes.

fn is_bond(&self, i: usize, j: usize) -> Result<bool, Error>

Tell if the atoms at indexes i and j are bonded together

fn is_angle(&self, i: usize, j: usize, k: usize) -> Result<bool, Error>

Tell if the atoms at indexes i, j and k constitues an angle

fn is_dihedral(&self, i: usize, j: usize, k: usize, m: usize) -> Result<bool, Error>

Tell if the atoms at indexes i, j, k and m constitues a dihedral angle

fn bonds_count(&self) -> Result<usize, Error>

Get the number of bonds in the system

fn angles_count(&self) -> Result<usize, Error>

Get the number of angles in the system

fn dihedrals_count(&self) -> Result<usize, Error>

Get the number of dihedral angles in the system

fn bonds(&self) -> Result<Vec<[usize; 2]>, Error>

Get the list of bonds in the system

fn angles(&self) -> Result<Vec<[usize; 3]>, Error>

Get the list of angles in the system

fn dihedrals(&self) -> Result<Vec<[usize; 4]>, Error>

Get the list of dihedral angles in the system

fn add_bond(&mut self, i: usize, j: usize) -> Result<(), Error>

Add a bond between the atoms at indexes i and j in the system

fn remove_bond(&mut self, i: usize, j: usize) -> Result<(), Error>

Remove any existing bond between the atoms at indexes i and j in the system

unsafe fn from_ptr(ptr: *const CHFL_TOPOLOGY) -> Topology

Create a Topology from a C pointer. This function is unsafe because no validity check is made on the pointer.

unsafe fn as_ptr(&self) -> *const CHFL_TOPOLOGY

Get the underlying C pointer. This function is unsafe because no lifetime guarantee is made on the pointer.